serial: allow the device to be set to `null`
It's useful to be able to log the serial console output to a ring
buffer (or even ASL) without necessarily having a real host console
attached to it (i.e. no TTY or `stdio`).

The code requires an open file descriptor so add a new type `null`
which opens `/dev/null`.

Signed-off-by: David Scott <dave@recoil.org>